Fix parsing error on leading whitespace (bug #17390)
The basic problem is as follows:
There is an incoming message that looks like the following:
":jonner!n=jonner@jalfrezi.collabora.co.uk PRIVMSG #channelname : foo bar"

When this message gets tokenized, we get the following tokens:
 - :jonner!n=jonner@jalfrezi.collabora.co.uk
 - PRIVMSG
 - #channelname
 - :
 - foo
 - bar

When we get to parsing the chat message portion, our current token is simply
":".  There is a check in the code that was meant to prevent zero-length strings
from being added to the parsed arguments, but that check assumes that we'll only
be adding the current token to the parsed args.  In reality, we want to consume
the entire remaining string (which may contain multiple tokens) and add it to
args.  So instead of checking that the the character immediately after the ':'
is non-NULL in the strdup-ed token, we should really check that the character
immediately after the ':' is non-NULL in the original message string.

Bug was introduced in commit 02cf59fb7b7efebbec39422460ec13ae1e3f5c89