Improvements to twisted test IRC server
Now at least returns a WELCOME message if it detects a NICK and USER message is
sent by the client.  This allows the client to finally transition to the
CONNECTED state from CONNECTING, so it's progress.  The test still seems to hang
when attempting to disconnect, though, I'm not sure what is causing that yet.