Fix a warning about closing an invalid file descriptor
When running the new test suite for idle (not yet merged upstream), I was
getting tests aborted and created core files due to a warning about closing an
invalid file descriptor.  The actual source of the warning I was seeing was in
async_connect_data_destroy() while shutting down the IO channel.  I believe this
is because a few lines earlier, we close the IOChannel's file descriptor, then
glib attempts to close it again in the g_io_channel_shutdown() function.

We seem to be doing this many other places, but we don't always get warnings
about it because often the close() call comes after the g_io_channel_shutdown()
call, and we don't check the return status from close().  It seems to me that
since glib will free the file descriptor for us when we shut down the IOChannel,
there's not even any reason to carry around the fd separately in
_AsyncConnectData (since we can always get it from the IOChannel anyway), so I
removed that as well.

I haven't changed similar uses in the SSL server connection yet.