Fix the fd double-close in the ssl server connection class as well
As in the normal server connection class, there doesn't seem to be any need to
carry the fd around in the AsyncConnectData struct or in the class priv struct
since we can always retrieve it from the GIOChannel when necessary.  In
addition, g_io_channel_shutdown closes the fd, so there's no reason to close the
fd separately.  In fact, closing the fd separately will result in a glib warning
about closing an invalid file descriptor if the close() is done before the
channel shutdown.