call Disconnect() asynchronously to prevent false test failures
previously, the -success and the -success-ssl tests were failing because they
didn't get a dbus method return from the Disconnect() method.  I thought that
this was a problem with idle not returning the response, but it turns out it was
caused by the fact that the test called Disconnect() synchronously.  So while
the test was waiting for the dbus return, the QUIT handler could not run, so the
connection was not getting closed.  Apparently idle waits for the connection to
be closed before sending the dbus method return.

When I change the Disconnect() call to be async, the test irc client is allowed
to run and the tests now pass.