Only try to connect to the IRC server over TCP
Previously we were not specifying any socket type in our getaddrinfo call, so
our address results included both TCP and UDP candidates.  When connecting to
the TCP one failed, it would try the UDP one, which would lead to some
unexpected behavior (for instance, connect() on a UDP socket with O_NONBLOCK set
will return 0 rather than -1/EINPROGRESS, which violates some assumptions in the
code).  If the UDP socket did manage to get connected, it would violate other
assumptions (for instance, setting the TCP_NODELAY option doesn't make sense on
a non-TCP socket and triggers a warning).
So, it doesn't really make sense to connect to a IRC server with UDP anyway, so
this patch restricts telepathy-idle to use TCP sockets only