Handle connect()s that return success immediately
The documentation on connect() is cryptic about in what situations a
non-blocking socket can connect immediately, but in theory it is possible.  I
had several test failures caused by violating the assumption that the return
status from connect() is always -1 on a non-blocking socket.  These turned out
to be caused by attempting to connect a UDP socket (which apparently is
instantaneous even on non-blocking sockets), but it seems wise to handle this
case more gracefully anyway, rather than simply asserting.