Add documentation for IRC message spec
If you haven't worked with it in a little while, it can be a little challenging
to remember what the different characters mean for the IRC message spec.  This
just adds a small comment describing each of the format letters so that it's
easier to add new IRC message types in the future.