Free a bunch more memory leaks reported by valgrind
Many telepathy signals take GPtrArrays of specialized dbus-glib boxed types.
The basic pattern for all of these leaks is the same:
- create a GValue for the specialized dbus type
- call g_value_take_boxed(foo, specialized-dbus-type);
- dbus_g_type_struct_set(...) /* set struct props */
- g_ptr_array_add(array, g_value_get_boxed())
- emit tp signal
- free ptr array

The problem is that even though the pointer array of the boxed types are freed,
the boxed types themselves are not properly freed.  So you need to go through
the list and call g_boxed_free on all of the array elements before freeing the
array (or call g_value_unset() after you're done with the ptr array, which will
result in the boxed type being freed by the GValue since it was initialized with
take_boxed()).

In addition, there was a somewhat related/intertwined bug in the topic handlers:
- in idle_muc_channel_topic_touch(), where we were calling g_value_get_boxed()
  and adding it to the GPtrArray, then modifying the GValue, calling
  g_value_get_boxed(), and adding that to the array again.  The intent was to
  add two different objects to the array, but since we were not copying the
  boxed type (e.g. g_value_dup_boxed()), the result was that we were adding a
  pointer to a boxed type in the array, modifying the object and then adding it
  again, so we ended up with two identical boxed type objects in the array
  rather than two different objects.  I'm not entirely sure what effect this bug
  had on things, but it should be fixed now regardless.
- in idle_muc_channel_topic_full(), we had the same problem as above, but in
  addition, the 'subject' property was never actually added to the ptr array, so
  I added this call.