Add a couple tests for slightly buggy server behavior
gimpnet IRC server has a bug where if you QUIT while it's waiting for an ident
response (at 'login'), it will totally ignore the QUIT message and go on to send
you a welcome message. I've added a test which approximates this behavior.

In addition, if you send quit behavior, but the remote server doesn't close the
socket, telepathy-idle will never actually quit (or even respond to the
Disconnect() dbus method, in fact). So I've added a test for this case.

I think we should probably unilaterally close the connection and exit after a
certain time if the server is misbehaving in one of these ways