Fix Bug #19766: private messages no longer delivered to group chat
This could perhaps be a slightly controversial change.  The reason that the bug
happened was that telepathy-idle's room handle repository normalizer function
had some special code to make joining chatrooms more user-friendly.  Basically,
if the user specified a chatroom 'foo', it would assume you meant '#foo' and
automatically prepend the '#' character.
The way the parser currently works is that when we get in a privmsg, we try to
parse the 'destination' of the privmsg as both a contact and a room. The way we
determine if it's a valid room or not is by calling tp_handle_ensure(room_repo,
...) (and same for determining valid contacts).  Unfortunately, because of the
automatic #-prepending this means that a single privmsg destination can be both
a valid contact and a valid channel name, so if the user is in a channel with
the same name as their nick (but without the leading #), the privmsg will be
handled by both the IMFactory and the MUCFactory