Speed up the tests a bit
When I was figuring out the test framework and trying to get things to work, I
had bumped up the time out on a bunch of tests to 10s rather than the default 5.
This means that a timeout failure takes a lot longer to happen and there's
really no need for a full 10s timeout.  Also, when testing that messages are
sent in the right order, there's no need to send 10 messages, so cut a few of
them.  Since there's a built-in 2s delay between each sent message, that means
the test was always taking *at least* 20s to execute.