Fix a bug and add a test for requesting multiple roomlist channels
I hadn't really excercised the code for returning an existing roomlist channel
if we had already created one.  So I added a test for it and in the process
discovered a bug, which is also fixed by this patch.

(The compiler found this bug when it was introduced while I was rebasing
Jonathon's branch, so I fixed it there and then. -Will)