Expose the nick validation function and use it in the CM
Add a filter for the 'account' param of the RequestConnection method so that it
can check if the passed nickname is valid for an IRC nick and reject it if not.
Previously, telepathy-idle happily accepted invalid nicknames and then didn't
provide any decent feedback about what was wrong when the 'login' failed.