Validate nicks per the IRC RFC
Previously we were allowing the nick to consist of any unicode alpha or digit
characters, but the IRC spec (RFC2812 section 2.3.1) is very explicit about
which characters are valid for a nick, and it's only ascii chars.

In addition, we weren't disallowing '-' as the first character of a nickname as
required by the IRC spec.

Added a few additional tests for valid and invalid nicks