Add a test for improper splitting of long messages (Bug #17392)
See the bug for details, but essentially, telepathy-idle will internally split
long messages (those above the IRC max length of 512) and send them in multiple
messages.  The problem is that the server then has to relay them on to the other
members of the channel, and when it does this, it adds a ':nick!server.name here
' prefix onto the message.  So if our message was exactly 512 bytes, when this
prefix gets added, it needs to chop some of the message off the end to fit under
the 512-char limit.  It does so silently, unfortunately.  So we'll need to
either limit our max messages to a value considerably lower than 512 to leave
room for the server to add their prefix or we'll need to figure out a way to
calculate very closely what the prefix is going to be (for the longest possible
nickname)