Split long messages properly (Bug #17392)
Use the newly-added idle_connection_get_max_message_length() API to split long
messages so that they allow for the server to add a prefix onto the messages
that it relays to others instead of simply using 512 as the max message length

I had to modify the test a bit to get it to behave like a real IRC server, but
now the test passes fine, and I did quite a bit of manual testing (copy/pasting
large chunks of test between two accounts) and have not seen any issues yet.
Part of my manual testing included copy/pasting large amounts of unicode text
from various wikipedia articles in other languages, so I believe that the utf-8
splitting issue is also fixed properly in this branch.