messages/leading-space.py: simplify this test significantly
wjt questioned the need for the subclassed server in his review and eliminating
the server actually makes things much simpler.  At the same time, I realized
that there's really no need for the user to join a chat room in order to test
out this behavior, so remove that whole piece and just send a privmsg directly
and test whether it is parsed correctly.