Bug #17354 - Emits Send immediately, even if the message is queued
Contrary to what the bug implies, it was decided that it's OK to emit the 'Sent'
signal immediately even if the message is only queued and is not yet actually
sent.  But there is still a problem in that if the message was long enough to be
split into multiple messages, only one 'Sent' signal is emitted instead of one
for each separate message.

This patch emits a signal for each individual message.