Be strict for what nicks we allow, and liberal for what nicks we accept from others
Bip has a nasty habit of sending backlog activity from a fictional '-bip' user.
Unfortunately, a nick with a leading '-' character is illegal according to the
IRC RFCs, but we have to accept it since people appear to be using it in the
wild.

This patch adds a flag to the nick validation script that specifies whether we
should use strict mode or lenient mode.  We use strict mode for validating local
nicknames, and lenient mode for validating remote nicknames.