Define _GNU_SOURCE not __USE_GNU to get strnlen(3).
__USE_GNU is a glibc-internal macro that we should not be using; this
was causing a build failure in openSUSE, which has a glibc from the
future.

idle-connection.c doesn't actually use strnlen.

Based on a patch by Vincent Untz, which didn't compile for me: it seemed
like string.h was getting #included before the #define, and thus strnlen
wasn't exposed. Making the #define the first thing in idle-parser.c
fixed this.

Fixes: fd.o#22291