Fix a buffer overrun in channel name validation.
I can't find any coherent specification for this type of channel name
(which appear to start with ! and contain five upper-case or numeric
chars). All the RFCs just reference other RFCs but it seems to be
something to do with safe channels. Anyway... let's not blow up if
someone actually tries to use one of these with less than five
characters after the bang!