Handle RPL_WHOISHOST messages in response to RequestContactInfo
The RPL_WHOISHOST message indicates the host name and IP address from
which the nick is connected to the server. This message is received
only if the nick against which the WHOIS message was sent is the one
that sent the it.

This is an extension of RFC 2812 implemented by some IRC daemons. eg.,
ircd-seven (http://ur1.ca/3khrh), ShadowIRCd (http://ur1.ca/3khsr)
and UnrealIRCd (http://ur1.ca/43e4r).

It seems that some daemons also use the same numeric for
RPL_BANEXPIRED (http://www.alien.net.au/irc/irc2numerics.html) even
though I could not find an actual example. Therefore, to be safe, we
only consider a message to be RPL_WHOISHOST if the trailing part has
"is connecting from " as a prefix, followed by some non-whitespace
characters. However this will adversely affect some versions of
IRC-Hispano's daemon (http://ur1.ca/43e55), which uses Spanish text,
but we can live with that.

Fixes: https://bugs.freedesktop.org/34796