Handle RPL_WHOISLOGGEDIN messages in response to RequestContactInfo
The RPL_WHOISLOGGEDIN message indicates the account name of the nick
against which the WHOIS message was sent. On Freenode, a user's account
name and primary nick are one and the same.

This is an extension of RFC 2812 implemented by some IRC daemons. eg.,
ircd-seven (http://ur1.ca/3khrh) and ShadowIRCd (http://ur1.ca/3khsr).

Fixes: https://bugs.freedesktop.org/34796