idletest: return an empty RPL_WHOISCHANNELS in response to WHOIS
RFC 2812 defines RPL_WHOISCHANNELS as:
"<nick> :*( ( "@" / "+" ) <channel> " " )"
Therefore, there is the possibility of an empty message even though
preliminary testing with some IRC daemons (eg., Freenode, GIMPNet and
IRCNet) did not reveal such a scenario. Nevertheless to maintain
strict compliance with the RFC we should test for this as well.

Fixes: https://bugs.freedesktop.org/34796