Try not to mistake RPL_WHOWAS_TIME as RPL_WHOISLOGGEDIN
It seems that some daemons also use the same numeric for
RPL_WHOWAS_TIME (http://www.alien.net.au/irc/irc2numerics.html) even
though I could not find an actual example. Therefore, to be safe, we
only consider a message to be RPL_WHOISLOGGEDIN if the trailing part
is "is logged in as".

Fixes: https://bugs.freedesktop.org/34796