Connection: refactor adding queue timeout.
Previously, the two-second timeout for sending messages from the queue
was set up in two different places, and cleared in more. This is bad for
readability: also, it turned out that one place forgot to multiply the
number of seconds by a thousand!

This commit adds helper functions that do the right thing—including
using g_timeout_add_seconds for the sake of your battery.