MUCChannel: Implement Channel.Interface.Subject
This patch pulls the topic-tracking and -setting code out of the
Telepathy.Properties implementation, uses it to implement the Subject
interface, and stubs out the Telepathy.Properties implementation.
Because Subject uses DBus.Properties.PropertiesChanged, which clashes
with Telepathy.Properties.PropertiesChanged, we have to remove the
GInterface for Telepathy.Properties from the object.

The latter is not tested anywhere! So this gives us a commit that
compiles, runs and passes all the tests, without having to cut out all
the old code which we want to salvage for RoomConfig.