Connection: replace \r and \n with spaces
If a message gets as far as _send_with_priority and contains \r or \n,
it's almost certainly not what the user anticipated. For instance,
before this fix, calling

    RemoveMembers([self_handle], "bye\r\nJOIN #telepathy")

would cause the user to leave the channel with the message "bye", and
then accidentally join #telepathy.

Rather than trying to strip out \r and \n everywhere, this patch just
replaces them with spaces just before sending.

https://bugs.freedesktop.org/show_bug.cgi?id=34840
Reviewed-by: Debarshi Ray <rishi@gnu.org>