IdleConnection: Handle connecting -> disconnected correctly
In case an error is encountered while trying to connect, it is
reported using tp_base_connection_disconnect_with_dbus_error, which
changes the state of the parent TpBaseConnection from
TP_CONNECTION_STATUS_CONNECTING to TP_CONNECTION_STATUS_DISCONNECTED.
Hence this should not be done again in sconn_status_changed_cb.

It is to be noted that when the Disconnect method is called, the
parent's status is immediately set to disconnected, and when the
underlying network connection finally disconnects and emits
"status-changed" we still need to finish shutting down the connection
in the callback. This leads to a situation where it may appear that
the state is changing from disconnected to disconnected, but it is
actually not so.

Also, until the underlying network connection is successfully made,
priv->conn does not point to a valid IdleServerConnection. Therefore
the disconnect method should not be called on the underlying network
connection before that.

Fixes: https://bugs.freedesktop.org/37145