IdleConnection: Prevent overlapping and duplicate sends
If the timeout is low or the underlying network is bad, then successive
calls to idle_server_connection_send_async can overlap. ie. a newer
operation is started before all the earlier ones have finished. The
underlying GOutputStream does not allow this.

Therefore, we add a new private member to track whether we are in the
middle of sending something or not, and refuse to send anything new in
the current timeout interval if that is the case.

Similarly, the same message can be submitted more than once if we do
not remove from the queue those that have been passed to
idle_server_connection_send_async, but has not had their operations
finish. In case an error occurs, the messages are restored to the
front of the queue for re-submission in the next interval.

Fixes: https://bugs.freedesktop.org/37145