Connection: simplify adding queue timeouts
So previously there was this special-case in _send_with_priority()
which would send out the message immediately rather than enqueueing it
if:

• it was max priority, or
• we're connected, and there's no timeout, and it's been at least
  timeout seconds since we last sent something.

The “max priority doesn't need to wait for the queue” case is
dysfunctional: if you call idle_server_connection_send_async() while
another call is in flight, then it fails (because
g_output_stream_write_async() fails in that case), and so
_send_with_max_priority_ready() would just push the message to the top
of the queue where it belongs.

The “flush immediately” logic is much clearer if added to the code
which sets up the timeouts.

The “if we're connected” check didn't do anything anyway:
_add_msg_to_queue() adds a timeout regardless of whether we're connected
(which is good, because it gets used during connection…).