Connection: disconnect if we get no PONG for our PING
Previously we'd just send out a PING into the æther every n seconds, and
pay no attention to whether we ever got an answer. So it was perfectly
possible for the connection to just sit there until the TCP timeout
kicks in, which I think is a matter of hours by default.

With this patch, if we haven't heard a PONG 3 keepalive-intervals after
sending a PING, Idle throws its toys out of the pram.

This has been tested as follows:

* Put my laptop on a wired and wireless network simultaneously.
* Connect to an IRC server. (The wired network is used.)
* Pull the network cable out. Idle is too stupid to realise the link it
  was using is gone, and because we're still ostensibly online, nothing
  tells it to disconnect.
* Wait keepalive-interval * 4, and watch the connection get
  disconnected.

It works both with a direct connection to Freenode, and with a
connection over an SSH tunnel to irssi-proxy.