Connect to the server in the non-main thread
To do interactive certificate verification with GTls one needs to block
in the accept-certificate signal handler, which practically means the
connection needs to be done in a seperate check. As a first step,
instead using _connect_to_host_async in the main thread use a GTask to
synchronously connect in a different thread instead.