Do not part from channels in destroy handler
In case the client suddenly disappears (e.g. crashes) the user was
previously parted from the channels. This is terrible behaviour when the
user is using a bouncer, as they expect to stay in their channels. This
handler doesn't seem to be called on clean exit of clients, only when
they disappear suddenly.